Move comments to separate module
