+++ src/comments/comments.service.ts
@@ -2,10 +2,10 @@ import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 
-import { CreateCommentDto } from "../dto/create-comment.dto";
-import { UpdateCommentDto } from "../dto/update-comment.dto";
+import { CreateCommentDto } from "./dto/create-comment.dto";
+import { UpdateCommentDto } from "./dto/update-comment.dto";
 
-import { Comment } from "../entities/comment.entity";
+import { Comment } from "./comment.entity";
 
 @Injectable()
 export class CommentsService {
