+++ src/products/products.module.ts
@@ -6,14 +6,8 @@ import { ProductsService } from "./services/products.service";
 import { ProductsController } from "./controllers/products.controller";
 
 import { User } from "../users/entities/user.entity";
-
 import { Category } from "../categories/category.entity";
-import { CategoriesService } from "../categories/categories.service";
-import { CategoriesController } from "../categories/categories.controller";
-
-import { Comment } from "./entities/comment.entity";
-import { CommentsService } from "./services/comments.service";
-import { CommentsController } from "./controllers/comments.controller";
+import { Comment } from "../comments/comment.entity";
 
 import { Image } from "./entities/image.entity";
 import { ImagesService } from "./services/images.service";
@@ -22,23 +16,24 @@ import { RatingsService } from "./services/ratings.service";
 import { RatingsController } from "./controllers/ratings.controller";
 import { Rating } from "./entities/rating.entity";
 import { OrderRow } from "../orders/entities/order-row.entity";
+
 import { CategoriesModule } from "../categories/categories.module";
+import { CommentsModule } from "../comments/comments.module";
 
 @Module({
   providers: [
     ProductsService,
-    CommentsService,
     ImagesService,
     RatingsService,
   ],
   controllers: [
     ProductsController,
-    CommentsController,
     ImagesController,
     RatingsController,
   ],
   imports: [
     CategoriesModule,
+    CommentsModule,
     TypeOrmModule
     .forFeature([
       Product,
